+++ Common/Objects.lua
@@ -19,6 +19,18 @@
 _BattlePetCompletionist = {
     Constants = {
         PET_QUALITY_RARE = 4,
+        PET_TYPE_ICONS = {
+            "Interface\\icons\\Icon_PetFamily_Humanoid",
+            "Interface\\icons\\Icon_PetFamily_Dragon",
+            "Interface\\icons\\Icon_PetFamily_Flying",
+            "Interface\\icons\\Icon_PetFamily_Undead",
+            "Interface\\icons\\Icon_PetFamily_Critter",
+            "Interface\\icons\\Icon_PetFamily_Magical",
+            "Interface\\icons\\Icon_PetFamily_Elemental",
+            "Interface\\icons\\Icon_PetFamily_Beast",
+            "Interface\\icons\\Icon_PetFamily_Water",
+            "Interface\\icons\\Icon_PetFamily_Mechanical",
+        },
         PET_SOURCES = {
             [1] = BATTLE_PET_SOURCE_1,
             [2] = BATTLE_PET_SOURCE_2,
@@ -28,6 +40,21 @@ _BattlePetCompletionist = {
             [7] = BATTLE_PET_SOURCE_7,
             [9] = BATTLE_PET_SOURCE_9
         },
+        PET_SOURCE_ICONS = {
+            [BATTLE_PET_SOURCE_1]  = "Interface/WorldMap/TreasureChest_64",         -- Drop
+            [BATTLE_PET_SOURCE_2]  = "Interface/GossipFrame/AvailableQuestIcon",    -- Quest
+            [BATTLE_PET_SOURCE_3]  = "Interface/Minimap/Tracking/Banker",           -- Vendor
+            [BATTLE_PET_SOURCE_4]  = "Interface/Archeology/Arch-Icon-Marker",       -- Profession
+            [BATTLE_PET_SOURCE_5]  = "Interface/Icons/Tracking_WildPet",            -- Pet Battle
+            -- 6 Achievement; no icon assigned
+            [BATTLE_PET_SOURCE_7]  = "Interface/GossipFrame/DailyQuestIcon",        -- World Event
+            [BATTLE_PET_SOURCE_8]  = "Interface/Minimap/Tracking/Banker",           -- Promotion
+            [BATTLE_PET_SOURCE_9]  = "Interface/Icons/inv_misc_hearthstonecard_legendary", -- Trading Card Game
+            [BATTLE_PET_SOURCE_10] = "Interface/Icons/item_shop_giftbox01",         -- Shop
+            [BATTLE_PET_SOURCE_11] = "Interface/Icons/Garrison_Building_MageTower", -- Discovery
+            [BATTLE_PET_SOURCE_12] = "Interface/Icons/TradingPostCurrency",         -- Trading Post
+        },
+        PET_SOURCE_ICON_FALLBACK = "Interface/Icons/Inv_misc_questionmark",
     },
     Enums = {
         CombatMode = {